+++ src/com/amplitude/api/AmplitudeClient.java
@@ -129,18 +129,27 @@ public class AmplitudeClient {
     protected String deviceId;
     private boolean newDeviceIdPerInstall = false;
     private boolean useAdvertisingIdForDeviceId = false;
-    private boolean initialized = false;
+    protected boolean initialized = false;
     private boolean optOut = false;
     private boolean offline = false;
     private boolean logDiagnosticEvents = false;
     private boolean flushSessionEventsOnLog = false;
+
+    /**
+     * Event metadata
+     */
+    long sessionId = -1;
+    long sequenceNumber = 0;
+    long lastEventId = -1;
+    long lastIdentifyId = -1;
+    long lastEventTime = -1;
+    long previousSessionId = -1;
 
     private DeviceInfo deviceInfo;
 
     /**
      * The current session ID value.
      */
-    long sessionId = -1;
     private int eventUploadThreshold = Constants.EVENT_UPLOAD_THRESHOLD;
     private int eventUploadMaxBatchSize = Constants.EVENT_UPLOAD_MAX_BATCH_SIZE;
     private int eventMaxCount = Constants.EVENT_MAX_COUNT;
@@ -227,28 +236,43 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
-                if (!client.initialized) {
-                    AmplitudeClient.upgradePrefs(context);
-                    AmplitudeClient.upgradeSharedPrefsToDB(context);
-                    client.httpClient = new OkHttpClient();
-                    client.initializeDeviceInfo();
+                if (!initialized) {
+                    // this try block is idempotent, so it's safe to retry initialize if failed
+                    try {
+                        AmplitudeClient.upgradePrefs(context);
+                        AmplitudeClient.upgradeSharedPrefsToDB(context);
+                        httpClient = new OkHttpClient();
+                        initializeDeviceInfo();
 
-                    if (userId != null) {
-                        client.userId = userId;
-                        client.dbHelper.insertOrReplaceKeyValue(USER_ID_KEY, userId);
-                    } else {
-                        client.userId = client.dbHelper.getValue(USER_ID_KEY);
+                        if (userId != null) {
+                            client.userId = userId;
+                            dbHelper.insertOrReplaceKeyValue(USER_ID_KEY, userId);
+                        } else {
+                            client.userId = dbHelper.getValue(USER_ID_KEY);
+                        }
+                        Long optOutLong = dbHelper.getLongValue(OPT_OUT_KEY);
+                        optOut = optOutLong != null && optOutLong == 1;
+
+                        // try to restore previous session id
+                        previousSessionId = getLongvalue(PREVIOUS_SESSION_ID_KEY, -1);
+                        if (previousSessionId >= 0) {
+                            sessionId = previousSessionId;
+                        }
+
+                        // reload event meta data
+                        sequenceNumber = getLongvalue(SEQUENCE_NUMBER_KEY, 0);
+                        lastEventId = getLongvalue(LAST_EVENT_ID_KEY, -1);
+                        lastIdentifyId = getLongvalue(LAST_IDENTIFY_ID_KEY, -1);
+                        lastEventTime = getLongvalue(LAST_EVENT_TIME_KEY, -1);
+
+                        initialized = true;
+
+                    } catch (CursorWindowAllocationException e) {  // treat as uninitialized SDK
+                        logger.e(TAG, String.format(
+                           "Failed to initialize Amplitude SDK due to: %s", e.getMessage()
+                        ));
+                        client.apiKey = null;
                     }
-                    Long optOut = client.dbHelper.getLongValue(OPT_OUT_KEY);
-                    client.optOut = optOut != null && optOut == 1;
-
-                    // try to restore previous session id
-                    long previousSessionId = client.getPreviousSessionId();
-                    if (previousSessionId >= 0) {
-                        client.sessionId = previousSessionId;
-                    }
-
-                    client.initialized = true;
                 }
             }
         });
@@ -277,16 +301,11 @@ public class AmplitudeClient {
         return this;
     }
 
+    // this method should only be called from the background log thread
     private void initializeDeviceInfo() {
         deviceInfo = new DeviceInfo(context);
-        runOnLogThread(new Runnable() {
-
-            @Override
-            public void run() {
-                deviceId = initializeDeviceId();
-                deviceInfo.prefetch();
-            }
-        });
+        deviceId = initializeDeviceId();
+        deviceInfo.prefetch();
     }
 
     /**
@@ -488,6 +507,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) { // in case initialization failed
+                    return;
+                }
                 client.optOut = optOut;
                 dbHelper.insertOrReplaceKeyLongValue(OPT_OUT_KEY, optOut ? 1L : 0L);
             }
@@ -816,6 +838,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) {  // in case initialization failed
+                    return;
+                }
                 logEvent(
                     eventType, copyEventProperties, apiProperties,
                     copyUserProperties, copyGroups, timestamp, outOfSession
@@ -899,9 +924,9 @@ public class AmplitudeClient {
 
             event.put("api_properties", apiProperties);
             event.put("event_properties", (eventProperties == null) ? new JSONObject()
-                    : truncate(eventProperties));
+                : truncate(eventProperties));
             event.put("user_properties", (userProperties == null) ? new JSONObject()
-                    : truncate(userProperties));
+                : truncate(userProperties));
             event.put("groups", (groups == null) ? new JSONObject() : truncate(groups));
         } catch (JSONException e) {
             logger.e(TAG, e.toString());
@@ -918,13 +943,12 @@ public class AmplitudeClient {
      * @return the event ID if succeeded, else -1
      */
     protected long saveEvent(String eventType, JSONObject event) {
-        long eventId;
         if (eventType.equals(Constants.IDENTIFY_EVENT)) {
-            eventId = dbHelper.addIdentify(event.toString());
-            setLastIdentifyId(eventId);
+            lastIdentifyId = dbHelper.addIdentify(event.toString());
+            setLastIdentifyId(lastIdentifyId);
         } else {
-            eventId = dbHelper.addEvent(event.toString());
-            setLastEventId(eventId);
+            lastEventId = dbHelper.addEvent(event.toString());
+            setLastEventId(lastEventId);
         }
 
         int numEventsToRemove = Math.min(
@@ -946,7 +970,7 @@ public class AmplitudeClient {
             updateServerLater(eventUploadPeriodMillis);
         }
 
-        return eventId;
+        return eventType.equals(Constants.IDENTIFY_EVENT) ? lastIdentifyId : lastEventId;
     }
 
     // fetches key from dbHelper longValueStore
@@ -962,19 +986,9 @@ public class AmplitudeClient {
      * @return the next sequence number
      */
     long getNextSequenceNumber() {
-        long sequenceNumber = getLongvalue(SEQUENCE_NUMBER_KEY, 0);
         sequenceNumber++;
         dbHelper.insertOrReplaceKeyLongValue(SEQUENCE_NUMBER_KEY, sequenceNumber);
         return sequenceNumber;
-    }
-
-    /**
-     * Internal method to get the last event time.
-     *
-     * @return the last event time
-     */
-    long getLastEventTime() {
-        return getLongvalue(LAST_EVENT_TIME_KEY, -1);
     }
 
     /**
@@ -983,16 +997,8 @@ public class AmplitudeClient {
      * @param timestamp the timestamp
      */
     void setLastEventTime(long timestamp) {
+        lastEventTime = timestamp;
         dbHelper.insertOrReplaceKeyLongValue(LAST_EVENT_TIME_KEY, timestamp);
-    }
-
-    /**
-     * Internal method to get the last event id.
-     *
-     * @return the last event id
-     */
-    long getLastEventId() {
-        return getLongvalue(LAST_EVENT_ID_KEY, -1);
     }
 
     /**
@@ -1001,16 +1007,8 @@ public class AmplitudeClient {
      * @param eventId the event id
      */
     void setLastEventId(long eventId) {
+        lastEventId = eventId;
         dbHelper.insertOrReplaceKeyLongValue(LAST_EVENT_ID_KEY, eventId);
-    }
-
-    /**
-     * Internal method to get the last identify id.
-     *
-     * @return the last identify id
-     */
-    long getLastIdentifyId() {
-        return getLongvalue(LAST_IDENTIFY_ID_KEY, -1);
     }
 
     /**
@@ -1019,6 +1017,7 @@ public class AmplitudeClient {
      * @param identifyId the identify id
      */
     void setLastIdentifyId(long identifyId) {
+        lastIdentifyId = identifyId;
         dbHelper.insertOrReplaceKeyLongValue(LAST_IDENTIFY_ID_KEY, identifyId);
     }
 
@@ -1032,20 +1031,12 @@ public class AmplitudeClient {
     }
 
     /**
-     * Internal method to get the previous session id.
-     *
-     * @return the previous session id
-     */
-    long getPreviousSessionId() {
-        return getLongvalue(PREVIOUS_SESSION_ID_KEY, -1);
-    }
-
-    /**
      * Internal method to set the previous session id.
      *
      * @param timestamp the timestamp
      */
     void setPreviousSessionId(long timestamp) {
+        previousSessionId = timestamp;
         dbHelper.insertOrReplaceKeyLongValue(PREVIOUS_SESSION_ID_KEY, timestamp);
     }
 
@@ -1069,7 +1060,6 @@ public class AmplitudeClient {
 
         // no current session - check for previous session
         if (isWithinMinTimeBetweenSessions(timestamp)) {
-            long previousSessionId = getPreviousSessionId();
             if (previousSessionId == -1) {
                 startNewSession(timestamp);
                 return true;
@@ -1107,7 +1097,6 @@ public class AmplitudeClient {
     }
 
     private boolean isWithinMinTimeBetweenSessions(long timestamp) {
-        long lastEventTime = getLastEventTime();
         long sessionLimit = usingForegroundTracking ?
                 minTimeBetweenSessionsMillis : sessionTimeoutMillis;
         return (timestamp - lastEventTime) < sessionLimit;
@@ -1147,8 +1136,7 @@ public class AmplitudeClient {
             return;
         }
 
-        long timestamp = getLastEventTime();
-        logEvent(sessionEvent, null, apiProperties, null, null, timestamp, false);
+        logEvent(sessionEvent, null, apiProperties, null, null, lastEventTime, false);
     }
 
     /**
@@ -1160,6 +1148,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) {
+                    return;
+                }
                 refreshSessionTime(timestamp);
                 inForeground = false;
                 if (flushEventsOnClose) {
@@ -1178,6 +1169,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) {
+                    return;
+                }
                 startNewSessionIfNeeded(timestamp);
                 inForeground = true;
             }
@@ -1297,6 +1291,10 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) {  // in case initialization failed
+                    return;
+                }
+
                 // Create deep copy to try and prevent ConcurrentModificationException
                 JSONObject copy;
                 try {
@@ -1306,12 +1304,18 @@ public class AmplitudeClient {
                     return; // could not create copy
                 }
 
+                // sanitize and truncate properties before trying to convert to identify
+                JSONObject sanitized = truncate(copy);
+                if (sanitized.length() == 0) {
+                    return;
+                }
+
                 Identify identify = new Identify();
-                Iterator<?> keys = copy.keys();
+                Iterator<?> keys = sanitized.keys();
                 while (keys.hasNext()) {
                     String key = (String) keys.next();
                     try {
-                        identify.setUserProperty(key, copy.get(key));
+                        identify.setUserProperty(key, sanitized.get(key));
                     } catch (JSONException e) {
                         logger.e(TAG, e.toString());
                     }
@@ -1377,13 +1381,19 @@ public class AmplitudeClient {
     /**
      * Truncate values in a JSON object. Any string values longer than 1024 characters will be
      * truncated to 1024 characters.
+     * Any dictionary with more than 1000 items will be ignored.
      *
      * @param object the object
      * @return the truncated JSON object
      */
     public JSONObject truncate(JSONObject object) {
         if (object == null) {
-            return null;
+            return new JSONObject();
+        }
+
+        if (object.length() > Constants.MAX_PROPERTY_KEYS) {
+            logger.w(TAG, "Warning: too many properties (more than 1000), ignoring");
+            return new JSONObject();
         }
 
         Iterator<?> keys = object.keys();
@@ -1421,7 +1431,7 @@ public class AmplitudeClient {
      */
     public JSONArray truncate(JSONArray array) throws JSONException {
         if (array == null) {
-            return null;
+            return new JSONArray();
         }
 
         for (int i = 0; i < array.length(); i++) {
@@ -1473,6 +1483,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(client.apiKey)) {  // in case initialization failed
+                    return;
+                }
                 client.userId = userId;
                 dbHelper.insertOrReplaceKeyValue(USER_ID_KEY, userId);
             }
@@ -1499,6 +1512,9 @@ public class AmplitudeClient {
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(client.apiKey)) {  // in case initialization failed
+                    return;
+                }
                 client.deviceId = deviceId;
                 dbHelper.insertOrReplaceKeyValue(DEVICE_ID_KEY, deviceId);
             }
@@ -1517,6 +1533,9 @@ public class AmplitudeClient {
         logThread.post(new Runnable() {
             @Override
             public void run() {
+                if (TextUtils.isEmpty(apiKey)) {  // in case initialization failed
+                    return;
+                }
                 updateServer();
             }
         });
@@ -1571,8 +1590,8 @@ public class AmplitudeClient {
             }
 
             try {
-                List<JSONObject> events = dbHelper.getEvents(getLastEventId(), batchSize);
-                List<JSONObject> identifys = dbHelper.getIdentifys(getLastIdentifyId(), batchSize);
+                List<JSONObject> events = dbHelper.getEvents(lastEventId, batchSize);
+                List<JSONObject> identifys = dbHelper.getIdentifys(lastIdentifyId, batchSize);
 
                 final Pair<Pair<Long, Long>, JSONArray> merged = mergeEventsAndIdentifys(
                         events, identifys, batchSize);
@@ -1594,6 +1613,14 @@ public class AmplitudeClient {
             } catch (JSONException e) {
                 uploadingCurrently.set(false);
                 logger.e(TAG, e.toString());
+
+            // handle CursorWindowAllocationException when fetching events, defer upload
+            } catch (CursorWindowAllocationException e) {
+                uploadingCurrently.set(false);
+                logger.e(TAG, String.format(
+                    "Caught Cursor window exception during event upload, deferring upload: %s",
+                    e.getMessage()
+                ));
             }
         }
     }